% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_cluster}
\alias{calc_cluster}
\title{Cluster flowSet according to channels}
\usage{
calc_cluster(flow, panel, channels, clusters = 20)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{panel}{value from \code{\link{import_fcs_path}}}

\item{channels}{channels to cluster}

\item{clusters}{optional: number of metaclusters}
}
\value{
clustered flowSet
}
\description{
Cluster flowSet according to channels
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))

cluster <- calc_cluster(flow_item$data, flow_item$panel, flow_item$panel$antigen)
}
