% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calc_contrast}
\alias{calc_contrast}
\title{Calculate contrast}
\usage{
calc_contrast(summary, others)
}
\arguments{
\item{summary}{output of \code{\link{calc_summary}}}

\item{others}{other channels to consider}
}
\value{
tibble: contrast
}
\description{
Calculate contrast
}
\examples{
\dontrun{
TODO
}
}
