% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calc_glmmadmb}
\alias{calc_glmmadmb}
\title{Compute GLMM ADMB}
\usage{
calc_glmmadmb(metaclust, files_inc, selected, minimum = 50)
}
\arguments{
\item{metaclust}{annotated metaclusters}

\item{files_inc}{output of \code{\link{extract_files}}}

\item{selected}{selected metaclustering}

\item{minimum}{optional: minimum cells per cluster}
}
\value{
tibble: GLMM ADMB results
}
\description{
Compute GLMM ADMB
}
\examples{
\dontrun{
TODO
}
}
