% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_mem}
\alias{calc_mem}
\title{Perform marker enrichment modeling}
\usage{
calc_mem(expr_anno)
}
\arguments{
\item{expr_anno}{output of \code{\link{collate_expressions}}}
}
\value{
tibble: MEM values
}
\description{
Perform marker enrichment modeling
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)
expr_anno <- collate_expressions(flow_expr, exp_label)

mem <- calc_mem(expr_anno)
}
