% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_sim_hilbert}
\alias{calc_sim_hilbert}
\title{Calculate and plot Hilbert Similarity}
\usage{
calc_sim_hilbert(
  flow_exps,
  exp_label,
  files_inc,
  channels,
  bins = 10,
  dimension = 4
)
}
\arguments{
\item{flow_exps}{output of \code{\link{scale_expressions}}}

\item{exp_label}{output of \code{\link{extract_labels}}}

\item{files_inc}{value from \code{\link{extract_files}}}

\item{channels}{channels of interest}

\item{bins}{optional: number of bins}

\item{dimension}{optional: number of dimensions (> 2)}
}
\value{
\code{hilbert}: tibble: Hilbert similarities

\code{dists}: distance matrix
}
\description{
Calculate and plot Hilbert Similarity
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
flow_exps <- scale_expressions(flow_expr)
exp_label <- extract_labels(flow_item$data)
files_inc <- extract_files(exp_label)

sim_hilbert <- calc_sim_hilbert(flow_exps, exp_label, files_inc$files, flow_item$panel$antigen)
}
