% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_som}
\alias{calc_som}
\title{Run FlowSOM on flowSet}
\usage{
calc_som(flow, channels, clusters = 20)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{channels}{channels to consider}

\item{clusters}{optional: number of clusters to generate}
}
\value{
FlowSOM output
}
\description{
Run FlowSOM on flowSet
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))

som <- calc_som(flow_item$data, flow_item$panel$fcs_colname)
}
