% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_summary}
\alias{calc_summary}
\title{Summarize expression data for each channel}
\usage{
calc_summary(expr_anno, grouping)
}
\arguments{
\item{expr_anno}{output of \code{\link{collate_expressions}}}

\item{grouping}{columns to group by}
}
\value{
tibble: summarized expression data
}
\description{
Summarize expression data for each channel
}
\examples{
\dontrun{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)
expr_anno <- collate_expressions(flow_expr, exp_label)

summary <- calc_summary(expr_anno, c("channel", "strain", "time", "file"))
}
}
