% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{collate_expressions}
\alias{collate_expressions}
\title{Extract expression information from parsed FCS data}
\usage{
collate_expressions(flow_item, cluster = NULL, min_cells = 0)
}
\arguments{
\item{flow_item}{output of \code{\link{import_fcs_path}} or \code{\link{rescale_item}}}

\item{cluster}{optional: value from \code{\link{do_cluster}}}

\item{min_cells}{optional: minimum number of cells per file}
}
\value{
expression information
}
\description{
Extract expression information from parsed FCS data
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item, flow_item$panel$antigen)

xprc_info <- collate_expressions(flow_item, cluster$sce)
}
