% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{compute_glmm}
\alias{compute_glmm}
\title{Compute GLMM ADMB}
\usage{
compute_glmm(expn_anno, meta, min_cluster_cells = 0)
}
\arguments{
\item{expn_anno}{output of \code{\link{collate_expressions}}}

\item{meta}{"meta{n}": selected metaclustering method}

\item{min_cluster_cells}{optional integer: minimum cells per cluster}
}
\value{
list: GLMM ADMB results per cluster
}
\description{
Compute GLMM ADMB
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
exp_label <- extract_labels(flow_item$data)
cluster <- do_cluster(flow_item$data, flow_item$panel, flow_item$panel$antigen)
expn_anno <- collate_expressions(NULL, exp_label, cluster)

glmm_list <- compute_glmm(expn_anno, "meta5")
}
