% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{compute_lmer}
\alias{compute_lmer}
\title{Calculate lmer}
\usage{
compute_lmer(expr_anno, meta, min_files = 7)
}
\arguments{
\item{expr_anno}{output of \code{\link{collate_expressions}}}

\item{meta}{"meta{n}": selected metaclustering method}

\item{min_files}{optional: minimum files per cluster}
}
\value{
list: lmer results per cluster per channel
}
\description{
Calculate lmer
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)
cluster <- do_cluster(flow_item$data, flow_item$panel, flow_item$panel$antigen)
expr_anno <- collate_expressions(flow_expr, exp_label, cluster)

lmer_list <- compute_lmer(expr_anno, "meta6")
}
