% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{compute_results}
\alias{compute_results}
\title{Collate computation results}
\usage{
compute_results(results)
}
\arguments{
\item{results}{output of \code{\link{compute_glmm}} or \code{\link{compute_lmer}}}
}
\value{
tibble: results
}
\description{
Collate computation results
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item, flow_item$panel$antigen)
xprc_info <- collate_expressions(flow_item, cluster$sce)

glmm_list <- compute_glmm(xprc_info, "meta6")
lmer_list <- compute_lmer(xprc_info, "meta6")

glmm_data <- compute_results(glmm_list)
lmer_data <- compute_results(lmer_list)
}
