% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{create_expression_matrix}
\alias{create_expression_matrix}
\title{Collate expression data for each channel}
\usage{
create_expression_matrix(flow, panel, range = c(0, 1))
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{panel}{value from \code{\link{import_fcs_path}}}

\item{range}{optional: quantile range for scaling}
}
\value{
\code{matrix} tibble: expression matrix

\code{scaled} tibble: scaled expression matrix
}
\description{
Collate expression data for each channel
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
expression <- create_expression_matrix(flow_item$data, flow_item$panel)
}
