% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{create_meta_freq}
\alias{create_meta_freq}
\title{Repeat metadata values based on cell count}
\usage{
create_meta_freq(flow)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}
}
\value{
tibble: metadata values
}
\description{
Constructs a list based on metadata columns, filled with individual metadata
values multiplied by cell count.
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
meta_freq <- create_meta_freq(flow_item$data)
}
