% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{do_cluster}
\alias{do_cluster}
\title{Cluster flowSet according to channels}
\usage{
do_cluster(flow, panel, channels, grid_x = 10, grid_y = 10, clusters = 20)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{panel}{value from \code{\link{import_fcs_path}}}

\item{channels}{character: channels to consider}

\item{grid_x}{optional integer: number of columns of self-organising map}

\item{grid_y}{optional integer: number of rows of self-organising map}

\item{clusters}{optional integer: number of clusters to generate}
}
\value{
SingleCellExperiment with added cluster information
}
\description{
Cluster flowSet according to channels
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item$data, flow_item$panel, flow_item$panel$antigen)
}
