% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{do_som}
\alias{do_som}
\title{Run FlowSOM on flowSet}
\usage{
do_som(flow, channels, grid_x = 10, grid_y = 10, clusters = 20)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{channels}{character: channels to consider}

\item{grid_x}{optional integer: number of columns of self-organising map}

\item{grid_y}{optional integer: number of rows of self-organising map}

\item{clusters}{optional integer: number of clusters to generate}
}
\value{
see \code{\link[FlowSOM]{FlowSOM}}
}
\description{
Run FlowSOM on flowSet
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
som <- do_som(flow_item$data, flow_item$panel$fcs_colname)
}
