% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_expressions}
\alias{extract_expressions}
\title{Extract expression data from flowSet}
\usage{
extract_expressions(flow, panel)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}

\item{panel}{value from \code{\link{import_fcs_path}}}
}
\value{
tibble: expression matrix
}
\description{
Extract expression data from flowSet
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
}
