% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_files}
\alias{extract_files}
\title{Filter and plot experiment-file pairs based on cell count}
\usage{
extract_files(exp_label, minimum = 0, colour = "Strain")
}
\arguments{
\item{exp_label}{output of \code{\link{extract_expressions}}}

\item{minimum}{optional: minimum number of cells per experiment-file pair}

\item{colour}{optional: bar colour}
}
\value{
\code{file} vector: file names

\code{plot} ggplot2: plot
}
\description{
Filter and plot experiment-file pairs based on cell count
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
exp_label <- extract_labels(flow_item$data)
files_inc <- extract_files(exp_label)
}
