% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract metadata labels for expression data from flowSet}
\usage{
extract_labels(flow)
}
\arguments{
\item{flow}{value from \code{\link{import_fcs_path}}}
}
\value{
tibble: metadata labels
}
\description{
Extract metadata labels for expression data from flowSet
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
exp_label <- extract_labels(flow_item$data)
}
