% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{filter_files}
\alias{filter_files}
\title{Filter and plot experiment-file pairs based on cell count}
\usage{
filter_files(meta_freq, minimum)
}
\arguments{
\item{meta_freq}{output of \code{\link{create_meta_freq}}}

\item{minimum}{minimum number of cells per experiment-file pair}
}
\value{
\code{file} vector: file names
\code{plot} ggplot2: plot
}
\description{
Filter and plot experiment-file pairs based on cell count
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
meta_freq <- create_meta_freq(flow_item)
files_inc <- filter_files(meta_freq, 5000)
}
