% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{flow_calc_mem}
\alias{flow_calc_mem}
\title{Perform marker enrichment modeling}
\usage{
flow_calc_mem(matrix, files_inc, meta_freq, channel, downsample = 1, seed = 1)
}
\arguments{
\item{matrix}{value from \code{\link{create_expression_matrix}}}

\item{files_inc}{value from \code{\link{filter_files}}}

\item{meta_freq}{value from \code{\link{create_meta_freq}}}

\item{channel}{selected channel}

\item{downsample}{double: [0, 1) down-sample data}
}
\value{
MEM results
}
\description{
TODO: unsure what this function does, seek clarification
}
\examples{
\dontrun{0}
}
