% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{flow_FlowSOM}
\alias{flow_FlowSOM}
\title{Title}
\usage{
flow_FlowSOM(
  live.mat,
  files.include,
  som,
  flow_annos,
  channels,
  metas,
  clusters
)
}
\arguments{
\item{clusters}{}
}
\value{

}
\description{
Title
}
