% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{flow_proj_plot}
\alias{flow_proj_plot}
\title{Title}
\usage{
flow_proj_plot(
  flow_annos,
  live_scaled,
  pheno_S,
  colour = "Strain",
  facet = "Strain",
  clusts = NULL,
  metas = NULL
)
}
\arguments{
\item{metas}{}
}
\value{

}
\description{
Title
}
