% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{flow_similarity}
\alias{flow_similarity}
\title{Title}
\usage{
flow_similarity(
  flow_anno,
  live_scaled,
  files_filt,
  channels,
  colour = "Strain",
  bins = 11
)
}
\arguments{
\item{bins}{}
}
\value{

}
\description{
Title
}
