% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{flow_start_flowSOM}
\alias{flow_start_flowSOM}
\title{Title}
\usage{
flow_start_flowSOM(
  liveSet,
  functional,
  phenotypic,
  flow_type = "phenotypic",
  gridSize = 10,
  nClus = 6
)
}
\arguments{
\item{nClus}{}
}
\value{

}
\description{
Title
}
