% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_fcs_path}
\alias{import_fcs_path}
\title{Import FCS data from FCS folder}
\usage{
import_fcs_path(fcs_path, marker_index = 1)
}
\arguments{
\item{fcs_path}{path: path to FCS folder}
}
\value{
parsed FCS data
}
\description{
A FCS folder is a folder with a `fcs` sub-directory containing `.fcs` files,
optionally along with `labels.csv` (with headings `File+Name` and
`Experiment`) and `markers.csv` (with `marker_label` and `marker_name`).
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
}
