% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_fcs_root}
\alias{import_fcs_root}
\title{Import all FCS data from FCS folder sub-directories}
\usage{
import_fcs_root(fcs_root)
}
\arguments{
\item{fcs_root}{path: root directory of FCS folders}
}
\value{
list: parsed FCS data
}
\description{
A FCS folder is a folder with a `fcs` sub-directory containing `.fcs` files,
optionally along with `labels.csv` (with headings `File+Name` and
`Experiment`) and `markers.csv` (with `marker_label` and `marker_name`).
}
\examples{
flow_data <- import_fcs_root(system.file("extdata", package = "WebCytoMetry"))
}
