% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{plot_radviz}
\alias{plot_radviz}
\title{Project expressions into a two-dimensional circular plot}
\usage{
plot_radviz(exps_anno, channels)
}
\arguments{
\item{exps_anno}{output of \code{\link{collate_expressions}}}

\item{channels}{character: selected channels}
}
\value{
see \code{\link[Radviz]{do.radviz}}
}
\description{
Project expressions into a two-dimensional circular plot
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
flow_exps <- scale_expressions(flow_expr, c(0, 1))
exp_label <- extract_labels(flow_item$data)
exps_anno <- collate_expressions(flow_exps, exp_label)

plot_radviz <- plot_radviz(exps_anno, flow_item$panel$antigen)
plot(plot_radviz)
}
