% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prep_radviz}
\alias{prep_radviz}
\title{Project expressions into a two-dimensional circular plot}
\usage{
prep_radviz(xpsc_info, channels)
}
\arguments{
\item{xpsc_info}{output of \code{\link{rescale_expressions}}}

\item{channels}{character: selected channels}
}
\value{
see \code{\link[Radviz]{do.radviz}}
}
\description{
Project expressions into a two-dimensional circular plot
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item, flow_item$panel$antigen)
xprc_info <- collate_expressions(flow_item, cluster$sce)
xpsc_info <- rescale_expressions(xprc_info)

radviz <- prep_radviz(xpsc_info, flow_item$panel$antigen)
plot(radviz)
}
