% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{rescale_expressions}
\alias{rescale_expressions}
\title{Scale expression information}
\usage{
rescale_expressions(expr_info, range = c(0, 1))
}
\arguments{
\item{expr_info}{output of \code{\link{collate_expressions}}}

\item{range}{optional [0, 1): quantile range}
}
\value{
scaled expression information
}
\description{
Scale expression information
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
expr_info <- collate_expressions(flow_item)
exps_info <- rescale_expressions(expr_info)
}
