% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{rescale_item}
\alias{rescale_item}
\title{Transform parsed FCS data with the arcsinh function}
\usage{
rescale_item(flow_item, ch_colnames = NULL, a = 1, b = 1, c = 0)
}
\arguments{
\item{flow_item}{output of \code{\link{import_fcs_path}}}

\item{ch_colnames}{character: channels to transform}

\item{a}{optional double: parameter a}

\item{b}{optional double: parameter b}

\item{c}{optional double: parameter b}
}
\value{
parsed FCS data
}
\description{
Expression data will be transformed with `asinh(a + b * x) + c`
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
flow_item <- rescale_item(flow_item, flow_item$panel$fcs_colname)
}
