% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{scale_expressions}
\alias{scale_expressions}
\title{Scale expression data}
\usage{
scale_expressions(flow_expr, range = c(0, 1))
}
\arguments{
\item{flow_expr}{output of \code{\link{extract_expressions}}}

\item{range}{optional: quantile range for scaling}
}
\value{
tibble: scaled expression matrix
}
\description{
Scale expression data
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
flow_exps <- scale_expressions(flow_expr)
}
