% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebCytoMetry-package.R
\docType{package}
\name{WebCytoMetry-package}
\alias{WebCytoMetry}
\alias{WebCytoMetry-package}
\title{WebCytoMetry: Web Interface for viewing flow cytometry data}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
