% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_mem}
\alias{calc_mem}
\title{Perform marker enrichment modeling}
\usage{
calc_mem(summary, downsample = 1, seed = 1)
}
\arguments{
\item{summary}{output of \code{\link{calc_summary}}}

\item{downsample}{[0, 1) optional: sampling factor}

\item{seed}{optional: random seed}
}
\value{
tibble: MEM values
}
\description{
Perform marker enrichment modeling
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)

summary <- calc_summary(flow_expr, exp_label, flow_item$panel$antigen[1])
mem <- calc_mem(summary)
}
