% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_sim_cosine}
\alias{calc_sim_cosine}
\title{Calculate cosine similarity}
\usage{
calc_sim_cosine(exps_info)
}
\arguments{
\item{exps_info}{output of \code{\link{rescale_expressions}}}
}
\value{
matrix: cosine similarities
}
\description{
Calculate cosine similarity
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
expr_info <- collate_expressions(flow_item)
exps_info <- rescale_expressions(expr_info)

sim_cosine <- calc_sim_cosine(exps_info)
}
