% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_sim_hilbert}
\alias{calc_sim_hilbert}
\title{Calculate Hilbert Similarity}
\usage{
calc_sim_hilbert(exps_info, channels, bins = 5, dimension = 4)
}
\arguments{
\item{exps_info}{output of \code{\link{collate_expressions}}}

\item{channels}{character: channels of interest}

\item{bins}{optional integer: number of bins}

\item{dimension}{optional integer: number of dimensions (> 2)}
}
\value{
\code{hilbert}: tibble: Hilbert similarities

\code{dists}: distance matrix
}
\description{
Calculate Hilbert Similarity
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
expr_info <- collate_expressions(flow_item)
exps_info <- rescale_expressions(expr_info)

sim_hilbert <- calc_sim_hilbert(exps_info, flow_item$panel$antigen)
}
