% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_summary}
\alias{calc_summary}
\title{Calculate summary}
\usage{
calc_summary(flow_expr, exp_label, selected)
}
\arguments{
\item{flow_expr}{output of \code{\link{extract_expressions}}}

\item{exp_label}{output of \code{\link{extract_labels}}}

\item{selected}{selected channel}
}
\value{
tibble: summary
}
\description{
Calculate summary
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)

summary <- calc_summary(flow_expr, exp_label, flow_item$panel$antigen[1])
}
