% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{collate_expressions}
\alias{collate_expressions}
\title{Attach expression labels to expression data, optionally with cluster data}
\usage{
collate_expressions(flow_expx, exp_label, cluster = NULL, min_file_cells = 0)
}
\arguments{
\item{flow_expx}{output of \code{\link{extract_expressions}} or \code{\link{scale_expressions}}}

\item{exp_label}{output of \code{\link{extract_labels}}}

\item{cluster}{optional: output of \code{\link{do_cluster}}}

\item{min_file_cells}{optional integer: minimum cells per file}
}
\value{
\code{data} tibble: annotated expression data

\code{channels} character: channels

\code{labels} character: labels
}
\description{
Attach expression labels to expression data, optionally with cluster data
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
exp_label <- extract_labels(flow_item$data)
cluster <- do_cluster(flow_item$data, flow_item$panel, flow_item$panel$antigen)

expr_anno <- collate_expressions(flow_expr, exp_label, cluster)
}
