% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{compute_glmm}
\alias{compute_glmm}
\title{Compute GLMM ADMB}
\usage{
compute_glmm(xprc_info, meta, min_cells = 0)
}
\arguments{
\item{xprc_info}{output of \code{\link{collate_expressions}}}

\item{meta}{"meta{n}": selected metaclustering method}

\item{min_cells}{optional integer: minimum cells per cluster}
}
\value{
list: GLMM ADMB results per cluster
}
\description{
Compute GLMM ADMB
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item, flow_item$panel$antigen)
xprc_info <- collate_expressions(flow_item, cluster$sce)

glmm_list <- compute_glmm(xprc_info, "meta6")
}
