% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{compute_lmer}
\alias{compute_lmer}
\title{Calculate lmer}
\usage{
compute_lmer(xprc_info, meta, formula, min_files = 7)
}
\arguments{
\item{xprc_info}{output of \code{\link{collate_expressions}}}

\item{meta}{"meta{n}": selected metaclustering method}

\item{formula}{character: formula expression}

\item{min_files}{optional: minimum files per cluster}
}
\value{
list: lmer results per cluster per channel
}
\description{
Calculate lmer
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "fcs", "KatJanin", package = "WebCytoMetry"))
cluster <- do_cluster(flow_item, flow_item$panel$antigen)
xprc_info <- collate_expressions(flow_item, cluster$sce)

lmer_list <- compute_lmer(xprc_info, "meta6", "median ~ strain + (1|strain)")
}
