% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_fcs_root}
\alias{import_fcs_root}
\title{Import all FCS data from FCS folder sub-directories}
\usage{
import_fcs_root(fcs_root)
}
\arguments{
\item{fcs_root}{path: root directory of FCS folders}
}
\value{
list: parsed FCS data
}
\description{
See \code{\link{import_fcs_path}} for a description of FCS folders
}
\examples{
flow_data <- import_fcs_root(system.file("extdata", "fcs", package = "WebCytoMetry"))
}
