% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{plot_radviz}
\alias{plot_radviz}
\title{Project expressions into a two-dimensional circular plot}
\usage{
plot_radviz(
  flow_exps,
  exp_label,
  sim_cosine,
  channels,
  colour = "experiment",
  facet = "experiment"
)
}
\arguments{
\item{flow_exps}{output of \code{\link{scale_expressions}}}

\item{exp_label}{output of \code{\link{extract_labels}}}

\item{sim_cosine}{output of \code{\link{calc_sim_cosine}}}

\item{channels}{selected channels}

\item{colour}{optional: dot color}

\item{facet}{optional: plot facet}
}
\value{
ggplot2: projection
}
\description{
Project expressions into a two-dimensional circular plot
}
\examples{
flow_item <- import_fcs_path(system.file("extdata", "KatJanin", package = "WebCytoMetry"))
flow_expr <- extract_expressions(flow_item$data, flow_item$panel)
flow_exps <- scale_expressions(flow_expr)
exp_label <- extract_labels(flow_item$data)
files_inc <- extract_files(exp_label)

sim_cosine <- calc_sim_cosine(flow_exps, exp_label, files_inc$files)
plot_radviz <- plot_radviz(flow_exps, exp_label, sim_cosine, flow_item$panel$antigen)
}
