% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebCytoMetry-package.R
\docType{package}
\name{WebCytoMetry-package}
\alias{WebCytoMetry}
\alias{WebCytoMetry-package}
\title{WebCytoMetry: Web Interface for viewing flow cytometry data}
\description{
Provides both command-line and web interfaces for visualising flow cytometry data.
}
\author{
\strong{Maintainer}: Lilin Yang \email{lilin.yang@uq.net.au}

Other contributors:
\itemize{
  \item Ahmed Mehdi \email{a.mehdi@imb.uq.edu.au} [contributor]
}

}
\keyword{internal}
